efi: Constify HandleProtocol() / LocateHandle() / LocateProtocol()
GUIDs passed to HandleProtocol() / LocateHandle() / LocateProtocol() are
not modified, so declare them const.  This allows the compiler to place
them in the rodata section instead of generating them on the stack at
runtime.  Pass GUIDs as literals rather than assigning them to temporary
variables to save a bit on code.  (Except for UGA and GOP GUIDs as they
are passed around between functions.)  Constification of the GUIDs is
in compliance with the EFI spec which marks them "IN" for said boot
services.

Cc: Matt Fleming <matt@codeblueprint.co.uk>
Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Lukas Wunner <lukas@wunner.de>