efi: Constify GetVariable() / SetVariable() signatures
GetVariable() / SetVariable() is typically called with immutable
VendorGuid and VariableName arguments.  SetVariable() is also often
called with an immutable Data argument.  However declaring these const
in the caller results in a compiler warning since they're not marked
const in the function signatures.  Rectify that, in keeping with the EFI
spec which marks these arguments "IN".

Cc: Matt Fleming <matt@codeblueprint.co.uk>
Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Cc: Tony Luck <tony.luck@intel.com>
Cc: Fenghua Yu <fenghua.yu@intel.com>
Cc: Boris Ostrovsky <boris.ostrovsky@oracle.com>
Cc: David Vrabel <david.vrabel@citrix.com>
Acked-by: Juergen Gross <jgross@suse.com>
Signed-off-by: Lukas Wunner <lukas@wunner.de>