efi: Constify GetVariable() / SetVariable() arguments
GUIDs and variable names passed to GetVariable() / SetVariable() are not
modified, so declare them const.  This allows the compiler to place them
in the rodata section instead of generating them on the stack at
runtime.  Pass GUIDs as literals rather than assigning them to temporary
variables to save a bit on code.

Cc: Matt Fleming <matt@codeblueprint.co.uk>
Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Cc: Tony Luck <tony.luck@intel.com>
Cc: Fenghua Yu <fenghua.yu@intel.com>
Cc: Anton Vorontsov <anton@enomsg.org>
Cc: Colin Cross <ccross@android.com>
Cc: Artur Paszkiewicz <artur.paszkiewicz@intel.com>
Cc: Mike Marciniszyn <mike.marciniszyn@intel.com>
Acked-by: Dennis Dalessandro <dennis.dalessandro@intel.com>
Reviewed-by: Kees Cook <keescook@chromium.org>
Signed-off-by: Lukas Wunner <lukas@wunner.de>