efi: Constify EFI_SIMPLE_TEXT_OUTPUT_PROTOCOL.OutputString()
UCS-2 strings passed to EFI_SIMPLE_TEXT_OUTPUT_PROTOCOL.OutputString()
are not modified, so can be declared const.  This allows the compiler to
place them in the rodata section instead of generating them on the stack
at runtime, such as with the '\r' string passed from efi_printk() to
efi_char16_printk().  Constification of such strings is in compliance
with the EFI spec which marks them "IN" for this protocol function.

Cc: Matt Fleming <matt@codeblueprint.co.uk>
Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Lukas Wunner <lukas@wunner.de>