efi: Constify EFI_RNG_PROTOCOL.GetRNG()
GUIDs passed to EFI_RNG_PROTOCOL.GetRNG() are not modified, so declare
them const.  This allows the compiler to place them in the rodata
section instead of generating them on the stack at runtime.  Pass GUIDs
as literals rather than assigning them to temporary variables to save a
bit on code.  Constification of the GUIDs is in compliance with the EFI
spec which marks them "IN" for this protocol function.

Cc: Matt Fleming <matt@codeblueprint.co.uk>
Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Lukas Wunner <lukas@wunner.de>